Refactored routes and controllers
