--- src/api/controllers/articles.controller.js
@@ -1,19 +1,0 @@
-import ArticleService from '../../business/services/article.service';
-import Response from './response';
-
-/**
- * ROUTE: articles/:username
- */
-
-const ArticlesController = {
-    get: (req, res) => {
-        const { username } = req.params;
-        ArticleService.getAllArticlesForAuthor(username, req.query).then((articles) => {
-            Response.json(res, 200, articles);
-        }).catch((err) => {
-            Response.json(res, 404, `No articles found, error was ${err}`);
-        });
-    }
-};
-
-export default ArticlesController;
