+++ src/api/routes/user.route.js
@@ -4,7 +4,7 @@ import bodyParser from 'body-parser';
 import { auth } from '../../config/jwt.config';
 
 import UserController from '../controllers/user.controller';
-import AuthController from '../controllers/auth.controller';
+import AuthGuard from '../guards/auth.guard';
 
 const router = express.Router();
 
@@ -15,8 +15,8 @@ router.use(bodyParser.urlencoded({
 // Availible via the base_url/user route
 router.route('/:username')
     .get(UserController.get)
-    .post(auth, AuthController.isAccessible, UserController.updateProfilePicture)
-    .put(auth, AuthController.isAccessible, UserController.put)
-    .delete(auth, AuthController.isAccessible, UserController.delete);
+    .post(auth, AuthGuard.isAccessible, UserController.updateProfilePicture)
+    .put(auth, AuthGuard.isAccessible, UserController.put)
+    .delete(auth, AuthGuard.isAccessible, UserController.delete);
 
 export default router;
